Make packages accessible by renaming internal to pkg
