+++ cmd/check_construction.go
@@ -20,8 +20,8 @@ import (
 	"os"
 	"time"
 
-	"github.com/coinbase/rosetta-cli/internal/tester"
-	"github.com/coinbase/rosetta-cli/internal/utils"
+	"github.com/slowboat0/rosetta-cli/pkg/tester"
+	"github.com/slowboat0/rosetta-cli/pkg/utils"
 
 	"github.com/coinbase/rosetta-sdk-go/fetcher"
 	"github.com/fatih/color"
