+++ cmd/utils_asserter_configuration.go
@@ -19,7 +19,7 @@ import (
 	"log"
 	"time"
 
-	"github.com/coinbase/rosetta-cli/internal/utils"
+	"github.com/slowboat0/rosetta-cli/pkg/utils"
 
 	"github.com/coinbase/rosetta-sdk-go/fetcher"
 	"github.com/spf13/cobra"
