+++ go.mod
@@ -1,20 +1,16 @@
-module github.com/coinbase/rosetta-cli
+module github.com/slowboat0/rosetta-cli
 
-go 1.13
+go 1.14
 
 require (
-	github.com/AndreasBriese/bbloom v0.0.0-20190825152654-46b345b51c96 // indirect
+	github.com/coinbase/rosetta-cli v0.4.0
 	github.com/coinbase/rosetta-sdk-go v0.3.3
 	github.com/dgraph-io/badger v1.6.1
 	github.com/ethereum/go-ethereum v1.9.18
 	github.com/fatih/color v1.9.0
-	github.com/golang/protobuf v1.4.2 // indirect
 	github.com/jinzhu/copier v0.0.0-20190924061706-b57f9002281a
-	github.com/pkg/errors v0.9.1 // indirect
 	github.com/spf13/cobra v1.0.0
-	github.com/spf13/pflag v1.0.5 // indirect
 	github.com/stretchr/testify v1.6.1
 	github.com/vmihailenco/msgpack/v5 v5.0.0-beta.1
 	golang.org/x/sync v0.0.0-20200317015054-43a5402ce75a
-	golang.org/x/sys v0.0.0-20200515095857-1151b9dac4a9 // indirect
 )
