+++ pkg/processor/constructor_handler.go
@@ -18,8 +18,8 @@ import (
 	"context"
 	"math/big"
 
-	"github.com/coinbase/rosetta-cli/internal/constructor"
-	"github.com/coinbase/rosetta-cli/internal/storage"
+	"github.com/slowboat0/rosetta-cli/pkg/constructor"
+	"github.com/slowboat0/rosetta-cli/pkg/storage"
 
 	"github.com/coinbase/rosetta-sdk-go/types"
 )
