+++ pkg/processor/reconciler_helper.go
@@ -18,7 +18,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/coinbase/rosetta-cli/internal/storage"
+	"github.com/slowboat0/rosetta-cli/pkg/storage"
 
 	"github.com/coinbase/rosetta-sdk-go/types"
 )
