+++ pkg/statefulsyncer/stateful_syncer.go
@@ -19,8 +19,8 @@ import (
 	"fmt"
 	"math/big"
 
-	"github.com/coinbase/rosetta-cli/internal/logger"
-	"github.com/coinbase/rosetta-cli/internal/storage"
+	"github.com/slowboat0/rosetta-cli/pkg/logger"
+	"github.com/slowboat0/rosetta-cli/pkg/storage"
 
 	"github.com/coinbase/rosetta-sdk-go/fetcher"
 	"github.com/coinbase/rosetta-sdk-go/syncer"
