+++ pkg/storage/block_storage.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/coinbase/rosetta-cli/internal/utils"
+	"github.com/slowboat0/rosetta-cli/pkg/utils"
 
 	"github.com/coinbase/rosetta-sdk-go/syncer"
 	"github.com/coinbase/rosetta-sdk-go/types"
