+++ pkg/storage/counter_storage.go
